Updated text fields pattern matching and MainApp

+++ RTS-C2-UD19/src/ex4/MiniCalculadora.java
@@ -15,7 +15,7 @@ public class MiniCalculadora {
 	{
 		switch (tipo) {
 			case Division:
-				resultado.setText(Integer.toString(operando1 / operando2));
+				resultado.setText(Double.toString((double)operando1 / operando2));
 				break;
 			case Multiplicacion:
 				resultado.setText(Integer.toString(operando1 * operando2));
